improvement(Confirmation): Confirmation is not a strategy.
