+++ lib/ash_authentication/dsl.ex
@@ -11,7 +11,7 @@ defmodule AshAuthentication.Dsl do
   alias Ash.{Api, Resource}
 
   alias AshAuthentication.{
-    Strategy.Confirmation,
+    AddOn.Confirmation,
     Strategy.OAuth2,
     Strategy.Password
   }
@@ -32,7 +32,8 @@ defmodule AshAuthentication.Dsl do
     ]
   ]
 
-  @default_lifetime_days 14
+  @default_token_lifetime_days 14
+  @default_confirmation_lifetime_days 3
 
   @secret_type {:or,
                 [
@@ -122,9 +123,9 @@ defmodule AshAuthentication.Dsl do
                 probably set this to a reasonably long time to ensure
                 a good user experience.
 
-                Defaults to #{@default_lifetime_days} days.
+                Defaults to #{@default_token_lifetime_days} days.
                 """,
-                default: @default_lifetime_days * 24
+                default: @default_token_lifetime_days * 24
               ],
               revocation_resource: [
                 type: {:behaviour, Resource},
@@ -144,7 +145,13 @@ defmodule AshAuthentication.Dsl do
             describe: "Configure authentication strategies on this resource",
             entities: [
               strategy(:password),
-              strategy(:oauth2),
+              strategy(:oauth2)
+            ]
+          },
+          %Section{
+            name: :add_ons,
+            describe: "Additional add-ons related to, but not providing authentication",
+            entities: [
               strategy(:confirmation)
             ]
           }
@@ -526,7 +533,103 @@ defmodule AshAuthentication.Dsl do
       name: :confirmation,
       describe: "User confirmation flow",
       target: Confirmation,
-      schema: Confirmation.schema()
+      schema: [
+        token_lifetime: [
+          type: :pos_integer,
+          doc: """
+          How long should the confirmation token be valid, in hours.
+
+          Defaults to #{@default_confirmation_lifetime_days} days.
+          """,
+          default: @default_confirmation_lifetime_days * 24
+        ],
+        monitor_fields: [
+          type: {:list, :atom},
+          doc: """
+          A list of fields to monitor for changes (eg `[:email, :phone_number]`).
+
+          The confirmation will only be sent when one of these fields are changed.
+          """,
+          required: true
+        ],
+        confirmed_at_field: [
+          type: :atom,
+          doc: """
+          The name of a field to store the time that the last confirmation took
+          place.
+
+          This attribute will be dynamically added to the resource if not already
+          present.
+          """,
+          default: :confirmed_at
+        ],
+        confirm_on_create?: [
+          type: :boolean,
+          doc: """
+          Generate and send a confirmation token when a new resource is created?
+
+          Will only trigger when a create action is executed _and_ one of the
+          monitored fields is being set.
+          """,
+          default: true
+        ],
+        confirm_on_update?: [
+          type: :boolean,
+          doc: """
+          Generate and send a confirmation token when a resource is changed?
+
+          Will only trigger when an update action is executed _and_ one of the
+          monitored fields is being set.
+          """,
+          default: true
+        ],
+        inhibit_updates?: [
+          type: :boolean,
+          doc: """
+          Wait until confirmation is received before actually changing a monitored
+          field?
+
+          If a change to a monitored field is detected, then the change is stored
+          in the token resource and  the changeset updated to not make the
+          requested change.  When the token is confirmed, the change will be
+          applied.
+
+          This could be potentially weird for your users, but useful in the case
+          of a user changing their email address or phone number where you want
+          to verify that the new contact details are reachable.
+          """,
+          default: true
+        ],
+        sender: [
+          type:
+            {:spark_function_behaviour, AshAuthentication.Sender,
+             {AshAuthentication.SenderFunction, 2}},
+          doc: """
+          How to send the confirmation instructions to the user.
+
+          Allows you to glue sending of confirmation instructions to
+          [swoosh](https://hex.pm/packages/swoosh),
+          [ex_twilio](https://hex.pm/packages/ex_twilio) or whatever notification
+          system is appropriate for your application.
+
+          Accepts a module, module and opts, or a function that takes a record,
+          reset token and options.
+
+          See `AshAuthentication.Sender` for more information.
+          """,
+          required: true
+        ],
+        confirm_action_name: [
+          type: :atom,
+          doc: """
+          The name of the action to use when performing confirmation.
+
+          If this action is not already present on the resource, it will be
+          created for you.
+          """,
+          default: :confirm
+        ]
+      ]
     }
   end
 end
