+++ lib/ash_authentication/info.ex
@@ -15,6 +15,7 @@ defmodule AshAuthentication.Info do
   def strategy(dsl_or_resource, name) do
     dsl_or_resource
     |> authentication_strategies()
+    |> Stream.concat(authentication_add_ons(dsl_or_resource))
     |> Enum.find_value(:error, fn strategy ->
       if strategy.name == name, do: {:ok, strategy}
     end)
