+++ test/test_helper.exs
@@ -1,7 +1,7 @@
+Mimic.copy(AshAuthentication.AddOn.Confirmation.Actions)
+Mimic.copy(AshAuthentication.AddOn.Confirmation.Plug)
 Mimic.copy(AshAuthentication.Plug.Defaults)
 Mimic.copy(AshAuthentication.Plug.Helpers)
-Mimic.copy(AshAuthentication.Strategy.Confirmation.Actions)
-Mimic.copy(AshAuthentication.Strategy.Confirmation.Plug)
 Mimic.copy(AshAuthentication.Strategy.OAuth2.Actions)
 Mimic.copy(AshAuthentication.Strategy.OAuth2.Plug)
 Mimic.copy(AshAuthentication.Strategy.Password.Actions)
